requirements.txt
altgraph==0.17
attrs==20.3.0
build==0.5.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorlog==5.0.1
coverage==5.5
docker==5.0.0
idna==2.10
iniconfig==1.1.1
macholib==1.14
packaging==20.9
pep517==0.10.0
pluggy==0.13.1
py==1.10.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
wcwidth==0.2.5
websocket-client==0.58.0
zipp==3.5.0
